refactor : Factory를 이용하는 방식으로만 파싱 전략 생성 가능하도록 변경
